Change if() in OptFunDome to suppress warning message*
Change if statement in OptFunDome so that "selectivityCurve" is
required to be just one input (%in%) of fixedFleetPars rather than THE
input (names(fixedFleetPars)== "selectivityCurve"). This bug arose from
the possibly redundant nature of the last logical expression in the if
clause
	if(length(tryFleetPars == 3) & length(fixedFleetPars) == 1 &
		c("selectivityCurve") %in% names(fixedFleetPars))
which might be better nested. Other adjustments are
- include nlminbOut in DoOptDome output, fix
- close sink(type = "message") in GTGDomeLBSPRSim by adding paired
  sink statement at end of function.

* warning message
Previously, when using the logistic curve with SL50 and SL95 specified
fixedFleetPars actually had three elements and was tested immediately
in the last expression of the if clause as
names(fixedFleetPars) == "selectivityCurve"

Warning messages:
1: In if (length(tryFleetPars) == 3 & length(fixedFleetPars) ==  ... :
  the condition has length > 1 and only the first element will be used